requirements.txt
arrow==0.15.4
autopep8==1.4.4
binaryornot==0.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.18.2
gevent==1.4.0
geventhttpclient==1.3.1
geventhttpclient-wheels==1.3.1.dev1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
locustio==0.12.1
MarkupSafe==1.1.1
msgpack-python==0.5.6
poyo==0.5.0
pycodestyle==2.5.0
python-dateutil==2.8.1
pyzmq==18.1.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
whichcraft==0.6.1
